adjust inheritance for mars/core and mars/serialize, serialize no longer inherit from mars/core, and SerializableWithKey and AttributeAsDictKey has been moved to mars/core
