+++ mars/serialize/__init__.py
@@ -16,7 +16,6 @@
 
 
 from .core import ValueType, Serializable, SerializableMetaclass, AttributeAsDict, \
-    SerializableWithKey, AttributeAsDictKey, \
     serializes, deserializes, ProviderType, Provider, \
     AnyField, IdentityField, BoolField, Int8Field, Int16Field, Int32Field, Int64Field, \
     UInt8Field, UInt16Field, UInt32Field, UInt64Field, Float16Field, Float32Field, Float64Field, \
